Split declaration modifiers into interface
- Create `HasModifiers` interface that associates a list of modifiers
  with a node
- Remove modifiers from `Declaration`
- Add `DeclarationWithModifiers` class for convenience
