+++ src/main/java/com/google/summit/ast/declaration/Declaration.kt
@@ -19,8 +19,7 @@ package com.google.summit.ast.declaration
 import com.google.summit.ast.Identifier
 import com.google.summit.ast.NodeWithSourceLocation
 import com.google.summit.ast.SourceLocation
-import com.google.summit.ast.modifier.AnnotationModifier
-import com.google.summit.ast.modifier.KeywordModifier
+import com.google.summit.ast.modifier.HasModifiers
 import com.google.summit.ast.modifier.Modifier
 
 /**
@@ -30,34 +29,6 @@ import com.google.summit.ast.modifier.Modifier
  * @param loc the location in the source file
  */
 sealed class Declaration(val id: Identifier, loc: SourceLocation) : NodeWithSourceLocation(loc) {
-
-  /**
-   * Constructs a declaration and sets modifiers.
-   *
-   * @param id the unqualified name of the class
-   * @param modifier the list of modifiers
-   * @param loc the location in the source file
-   */
-  constructor(id: Identifier, modifiers: List<Modifier>, loc: SourceLocation) : this(id, loc) {
-    this.modifiers = modifiers
-  }
-
-  /**
-   * The modifiers for this declaration.
-   *
-   * Because the modifiers are frequently parsed outside of the specific declaration rule in the
-   * Apex grammar, they are mutable and appended post-construction.
-   */
-  var modifiers: List<Modifier> = emptyList()
-
-  /** The subset of [modifiers] that are [annotations][AnnotationModifier]. */
-  val annotationModifiers
-    get() = modifiers.filterIsInstance<AnnotationModifier>()
-
-  /** The subset of [modifiers] that are [keywords][KeywordModifier]. */
-  val keywordModifiers
-    get() = modifiers.filterIsInstance<KeywordModifier>()
-
   /**
    * The qualified Name of the symbol includes any enclosing class(es) as prefixes, delimited by a
    * dot. For example: `OuterClass.InnerClass.innerMethod`
@@ -72,7 +43,33 @@ sealed class Declaration(val id: Identifier, loc: SourceLocation) : NodeWithSour
 
   /** Returns the enclosing type declaration. */
   fun getEnclosingType(): TypeDeclaration? = parent as? TypeDeclaration
+}
 
-  /** Returns whether this declaration has a [KeywordModifier] that matches the given [keyword]. */
-  fun hasKeyword(keyword: KeywordModifier.Keyword) = keywordModifiers.any { it.keyword == keyword }
+/**
+ * A [Declaration] that has associated [Modifier]s.
+ *
+ * @property id the unqualified name of the declaration
+ * @param loc the location in the source file
+ */
+sealed class DeclarationWithModifiers(id: Identifier, loc: SourceLocation) :
+  Declaration(id, loc), HasModifiers {
+
+  /**
+   * Constructs a declaration and sets modifiers.
+   *
+   * @param id the unqualified name of the declaration
+   * @param modifiers the list of modifiers
+   * @param loc the location in the source file
+   */
+  constructor(id: Identifier, modifiers: List<Modifier>, loc: SourceLocation) : this(id, loc) {
+    this.modifiers = modifiers
+  }
+
+  /**
+   * The modifiers for this declaration.
+   *
+   * Because the modifiers are frequently parsed outside of the specific declaration rule in the
+   * Apex grammar, they are mutable and appended post-construction.
+   */
+  final override var modifiers: List<Modifier> = emptyList()
 }
