+++ src/main/java/com/google/summit/ast/declaration/VariableDeclaration.kt
@@ -32,7 +32,7 @@ import com.google.summit.ast.modifier.Modifier
  * @property type a reference to the type of the variable
  * @property modifiers a list of any modifiers
  * @property initializer an optional initializer expression
- * @param sourceLocation the location in the source file
+ * @param loc the location in the source file
  */
 class VariableDeclaration(
   id: Identifier,
@@ -40,7 +40,7 @@ class VariableDeclaration(
   modifiers: List<Modifier>,
   val initializer: Expression?,
   loc: SourceLocation
-) : Declaration(id, modifiers, loc) {
+) : DeclarationWithModifiers(id, modifiers, loc) {
 
   override fun getChildren(): List<Node> = modifiers + listOfNotNull(id, type, initializer)
 }
